pyproject.toml
[project]
name = "mc3gs"
version = "0.1.0"
description = "Molecule-Constrained Gaussian Splatting for 3D scene reconstruction"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "MC-3GS Team" }
]
keywords = ["gaussian-splatting", "3d-reconstruction", "molecules", "neural-rendering"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Visualization",
]

dependencies = [
    "torch>=2.1.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "rdkit>=2023.9.1",
    "scipy>=1.11.0",
    "einops>=0.7.0",
    "plyfile>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "marimo>=0.18.4",
    "matplotlib>=3.10.8",
    "pyzmq>=27.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
]
# cuda = [
#     "diff-gaussian-rasterization>=0.0.0",
# ]

[project.scripts]
mc3gs-train = "mc3gs.cli.mc3gs_train:main"
mc3gs-export = "mc3gs.cli.mc3gs_export:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mc3gs"]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["mc3gs"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
